Start time warp algorithm

+++ Lab2/app/src/main/java/lab2_202_10/uwaterloo/ca/lab2/gesture/AccelerometerPoint.java
@@ -2,9 +2,9 @@ package lab2_202_10.uwaterloo.ca.lab2.gesture;
 
 public class AccelerometerPoint {
 
-    public int x, y, z;
+    public float x, y, z;
 
-    public AccelerometerPoint(int x, int y, int z) {
+    public AccelerometerPoint(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
